Add v2 Wikipedia data pipeline and experiment runner
- data/wikipedia.py: Full 6-step pipeline (download, embed, cluster,
  tokenize, build sequences) with per-step caching to disk
- run_experiment.py: Add --v2 flag for Wikipedia experiment (~98M param
  model, 30K steps) and --v2 --debug for quick testing
- data/synthetic.py: Fix int32/int64 dtype issue in dataset __getitem__
  (always return LongTensor for embedding layer compatibility)
- EXPERIMENT_LOG.md: Add sanity check results + v2 implementation details

+++ data/synthetic.py
@@ -251,8 +251,8 @@ class MicroLanguageDataset(Dataset):
 
     def __getitem__(self, idx: int) -> dict[str, torch.Tensor]:
         return {
-            "input_ids": torch.from_numpy(self.tokens[idx].copy()),
-            "labels": torch.from_numpy(self.labels[idx].copy()),
+            "input_ids": torch.from_numpy(self.tokens[idx].copy()).long(),
+            "labels": torch.from_numpy(self.labels[idx].copy()).long(),
         }
